moved Compression and Tools into their own project repos.
This is work in progress but I've copied the files into their own projects and deleted the originals. Now I just need to get them building again.

+++ Tests/TestLZSS.cpp
@@ -10,14 +10,11 @@ int main()
   std::random_device rd;
   std::mt19937 gen(rd());
   std::uniform_int_distribution<unsigned char> dis(0U);
-  const auto Test = [&dis,&gen](const size_t &size)
-  {
+  const auto Test = [&dis, &gen](const size_t &size) {
     if (size <= 0) { return true; }
-    std::vector<char> randomNums = std::vector<char>(static_cast<unsigned int>(size));
+    std::vector<char> randomNums = std::vector<char>(size);
     if (randomNums.empty()) { return true; }
-    std::generate(randomNums.begin(), randomNums.end(), [&dis,&gen]() {
-           return static_cast<char>(dis(gen));
-    });
+    std::generate(randomNums.begin(), randomNums.end(), [&dis, &gen]() { return static_cast<char>(dis(gen)); });
     auto compressed = OpenVIII::Compression::LZSS::Compress(randomNums);
     auto uncompressed = OpenVIII::Compression::LZSS::Decompress(compressed);
     if (std::equal(randomNums.begin(), randomNums.end(), uncompressed.begin())) {
@@ -30,7 +27,6 @@ int main()
 
   size_t i = 0;
   while (Test(i++)) {}
-
 
 
   return 0;
